chore: AppRootComponent 분리
